Merge libloragw
Before starting the retreat I merge libloragw into packet_forwarder
as a subdir to make the job easier. The coupling seems to be tight
anyways. I do this without trying to merge any history. I couldn't
google it out how to do it with the history in 5 minutes and I don't
think having the history adds that much value. This is good enough
for now. If we need some info from the history later, we can look
it up in the original repo which is this:

https://github.com/TheThingsNetwork/lora_gateway/tree/legacy